Remove WebHDFS::Factual::InnerClient
