+++ lib/webhdfs/factual/client.rb
@@ -4,13 +4,106 @@ module WebHDFS
       def initialize(api, default_namenode)
         @api = api
         @default_namenode = default_namenode
-        @client = get_client
+        @client = setup
       end
 
-      def get_client
-        InnerClient.setup(@api, @default_namenode)
+      private
+
+      def self.keytab_path_set?
+        if ENV['KEYTAB_PATH'].nil? || ENV['KEYTAB_PATH'].empty?
+          WebHDFS::Factual.logger.info("KEYTAB_PATH not set.")
+          false
+        else
+          true
+        end
       end
 
+      # FIXME: This should either not return false ever or should be handled
+      # FIXME: This should be renamed and its flow fixed
+      def setup
+        return false unless self.class.keytab_path_set?
+        @client = initialize_client
+        return false unless client_working?
+        @client
+      end
+
+      def initialize_client
+        @client = WebHDFS::Client.new(detect_namenode)
+        @client.kerberos = true
+        @client.kerberos_keytab = ENV['KEYTAB_PATH']
+        @client
+      end
+
+      def detect_namenode
+        conn = WebHDFS::Factual::APIConnection.new(@api)
+        path = 'jmx?qry=Hadoop:service=NameNode,name=NameNodeStatus'
+        res = conn.get(path)
+        namenode = res['beans'].first['HostAndPort'].split(':').first
+        namenode
+      rescue StandardError => e
+        WebHDFS::Factual.logger.warn("Failed to detect namenode with error: #{e}")
+        WebHDFS::Factual.logger.warn("Defaulting to #{@default_namenode}")
+        @default_namenode
+      end
+
+      def client_working?
+        res = @client.stat('/')
+        !!res
+      rescue StandardError => e
+        WebHDFS::Factual.logger.error("Failed to access HDFS with error #{$!}")
+        if e.is_a?(NameError) && e.message =~ /undefined local variable or method `min_stat'/
+          WebHDFS::Factual.logger.info("Do you have a valid keytab file at #{ENV['KEYTAB_PATH']}?")
+        end
+        raise e
+      end
+
+      def smart_retry(&block)
+        block.call
+      rescue WebHDFS::IOError => e
+        specific_exception = JSON.parse(e.message)['RemoteException']['exception'] rescue nil
+        message = JSON.parse(e.message)['RemoteException']['message'] rescue nil
+        if specific_exception == 'StandbyException'
+          WebHDFS::Factual.logger.error("HDFS namenode in standby. Sleeping for 10 seconds and then attempting to reconnect.")
+          Kernel.sleep 10
+          @client = setup
+          block.call
+        elsif message =~ /^Cannot obtain block length/
+          WebHDFS::Factual.logger.error(e.message)
+          Kernel.sleep 5
+          block.call
+        else
+          raise
+        end
+      rescue WebHDFS::ServerError => e
+        WebHDFS::Factual.logger.error(e.message)
+        Kernel.sleep 15
+        block.call
+      rescue WebHDFS::KerberosError => e
+        WebHDFS::Factual.logger.error("Kerberos credentials expired, refreshing them.")
+        @client = setup
+        block.call
+      end
+
+      def rescue_read_errors(path, &block)
+        block.call
+      rescue WebHDFS::FileNotFoundError => e
+        begin
+          message = JSON.parse(e.message)["RemoteException"]["message"]
+        rescue StandardError
+          message = e.message
+        end
+        if message =~ /not found/
+          # raise NeutronicHelper::FileNotFoundError, message, e.backtrace
+          raise WebHDFS::FileNotFoundError, message, e.backtrace
+        else
+          raise InvalidOpError, message, e.backtrace
+        end
+      end
+
+      public
+
+      # TODO
+      # The following are all HDFS methods.
       def append(path, data)
         smart_retry do
           begin
@@ -88,51 +181,6 @@ module WebHDFS
             raise WebHDFS::FileNotFoundError, "File #{path} not found", e.backtrace
           end
           Time.at(modification_time / 1000)
-        end
-      end
-
-      private
-
-      def smart_retry(&block)
-        block.call
-      rescue WebHDFS::IOError => e
-        specific_exception = JSON.parse(e.message)['RemoteException']['exception'] rescue nil
-        message = JSON.parse(e.message)['RemoteException']['message'] rescue nil
-        if specific_exception == 'StandbyException'
-          WebHDFS::Factual.logger.error("HDFS namenode in standby. Sleeping for 10 seconds and then attempting to reconnect.")
-          Kernel.sleep 10
-          @client = get_client
-          block.call
-        elsif message =~ /^Cannot obtain block length/
-          WebHDFS::Factual.logger.error(e.message)
-          Kernel.sleep 5
-          block.call
-        else
-          raise
-        end
-      rescue WebHDFS::ServerError => e
-        WebHDFS::Factual.logger.error(e.message)
-        Kernel.sleep 15
-        block.call
-      rescue WebHDFS::KerberosError => e
-        WebHDFS::Factual.logger.error("Kerberos credentials expired, refreshing them.")
-        @client = get_client
-        block.call
-      end
-
-      def rescue_read_errors(path, &block)
-        block.call
-      rescue WebHDFS::FileNotFoundError => e
-        begin
-          message = JSON.parse(e.message)["RemoteException"]["message"]
-        rescue StandardError
-          message = e.message
-        end
-        if message =~ /not found/
-          # raise NeutronicHelper::FileNotFoundError, message, e.backtrace
-          raise WebHDFS::FileNotFoundError, message, e.backtrace
-        else
-          raise InvalidOpError, message, e.backtrace
         end
       end
     end
