+++ spec/lib/webhdfs/factual/client_spec.rb
@@ -12,16 +12,31 @@ describe Client do
   it_behaves_like 'a target filesystem interface'
   it_behaves_like 'a target filesystem implementation'
 
+  describe '#detect_namenode' do
+    it 'returns the correct namenode' do
+      client = get_client
+      # TODO: This pattern should also accept API_HOST
+      stub_request(:get, /localhost/).to_return(body: fixture('namenode_status.json'))
+      expect(client.send(:detect_namenode)).to eq(DEFAULT_NAMENODE)
+    end
+
+    it 'returns the default namenode when request fails' do
+      client = get_client
+      stub_request(:get, /localhost/).to_return(status: 500)
+      expect(client.send(:detect_namenode)).to eq(DEFAULT_NAMENODE)
+    end
+  end
+
   context 'when hdfs namenode has changed' do
     before :each do
-      allow(InnerClient).to receive(:detect_namenode).and_return('bad_hdfs')
-      allow(InnerClient).to receive(:client_working?).and_return(true)
+      allow_any_instance_of(Client).to receive(:detect_namenode).and_return('bad_hdfs')
+      allow_any_instance_of(Client).to receive(:client_working?).and_return(true)
       @bad_hdfs = get_client
     end
 
     it 'switches to the correct namenode' do
-      allow(InnerClient).to receive(:detect_namenode).and_return(DEFAULT_NAMENODE)
-      allow(InnerClient).to receive(:client_working?).and_return(true)
+      allow_any_instance_of(Client).to receive(:detect_namenode).and_return(DEFAULT_NAMENODE)
+      allow_any_instance_of(Client).to receive(:client_working?).and_return(true)
       allow(Kernel).to receive(:sleep).and_return(true)
       expect {
         begin
